refactor: split backend CLI commands
